requirements.txt
colour==0.1.5
numpy==1.16.2
networkx==2.3
matplotlib==3.0.3
pylatexenc==1.4
antlr4-python3-runtime==4.7.2